Rename activities to inboxItems
